+++ Src/Library/Endpoint/Factory/EndpointFactory.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace FastEndpoints;
 
@@ -16,7 +15,7 @@ public class EndpointFactory : IEndpointFactory
     /// <param name="ctx">the http context for the current request</param>
     public BaseEndpoint Create(EndpointDefinition definition, HttpContext ctx)
     {
-        //note: if the default factory is being called, that means it's ok to use RequestServices below since the default MS DI is being used
+        //note: if the default factory is being called, that means it's ok to use HttpContext.RequestServices below since the default MS DI is being used
 
         var epInstance = (BaseEndpoint)Config.ServiceResolver.CreateInstance(definition.EndpointType, ctx.RequestServices);
 
@@ -24,7 +23,7 @@ public class EndpointFactory : IEndpointFactory
         {
             var prop = definition.ServiceBoundEpProps[i];
             prop.PropSetter ??= definition.EndpointType.SetterForProp(prop.PropName);
-            prop.PropSetter(epInstance, ctx.RequestServices.GetRequiredService(prop.PropType));
+            prop.PropSetter(epInstance, ctx.Resolve(prop.PropType));
         }
 
         return epInstance;