+++ Web/PipelineBehaviors/PreProcessors/MyRequestLogger.cs
@@ -6,7 +6,7 @@ public class MyRequestLogger<TRequest> : IPreProcessor<TRequest>
 {
     public Task PreProcessAsync(TRequest req, HttpContext ctx, List<ValidationFailure> failures, CancellationToken ct)
     {
-        var logger = ctx.RequestServices.GetRequiredService<ILogger<TRequest>>();
+        var logger = ctx.Resolve<ILogger<TRequest>>();
 
         logger.LogInformation($"request:{req?.GetType().FullName} path: {ctx.Request.Path}");